+++ src/commands/check.js
@@ -1,0 +1,4 @@
+// Command to check the enemy's remaining health
+export const check = (enemy) => {
+  return `The ${enemy.name} has ${enemy.hp} health remaining.`;
+};
