+++ src/commands/hunt.js
@@ -1,0 +1,5 @@
+import Dragon from '../enemies/Dragon.js';
+
+export const hunt = () => {
+  return new Dragon;
+};
